Add “big operator” support
The supported big operators are sum, product and a few similar.
Potentially, anything can be chosen to be a big operator and this can
easily be made a user setting (by making the `bigOperators` list
global), if we want to do that, although it might not be the cleanest
way.